Bodge test prediction scaling to work with different aspect ratios to kitti images.
It's a bodge as in the GT is used to pad the prediction to make the prediction the same aspect ratio as the test images.
It's sufficient to see that the training is converging, but the results aren't comparable to the original version.
Really the error should only be computed over the predict bit of the resulting resized prediction image.